requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.5
django-bootstrap==0.2.4
django-bootstrap3==9.0.0
django-bower==5.2.0
djangorestframework==3.7.1
gunicorn==19.7.1
idna==2.6
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
whitenoise==3.3.1
